Refactor theme
Separate out organisation's themes and choose which to use based on an env variable

Refs UserOfficeProject/stfc-user-office-project#69
