+++ src/styles/StyledComponents.tsx
@@ -4,7 +4,7 @@ import Paper from '@material-ui/core/Paper';
 import styled from '@material-ui/core/styles/styled';
 import React from 'react';
 
-import { getTheme } from '../theme';
+import { getTheme } from '../themes/theme';
 
 type PaddingMarginType = [number, number, number, number];
 type FlexDirectionType =
